Move adapters to lib module
